dia feliz consegui fazer bastante coisa da api me senti inteligente bora q tem mais

+++ SENAI_SP_Medical_Group/SENAI_SP_Medical_Group/Controllers/PacientesController.cs
@@ -47,7 +47,7 @@ namespace SENAI_SP_Medical_Group.Controllers
             }
         }
         [HttpPut("{id}")]
-        public IActionResult AtualizarUrl(short id, Paciente pacienteAtualizado)
+        public IActionResult AtualizarUrl(int id, Paciente pacienteAtualizado)
         {
             try
             {
